fix(cmd/list): fix `reading body EOF` error
We are using RPC to handle p2p messages. The RPC package encodes to gob, and gob has issues encoding more complicated structures. In this case, it cannot handle `DatasetPod.Structure.Schema`. To work around this, and since we dont need a schema to list a dataset, we remove the schema if we are using RPC